requirements.txt
mysql-connector-python-9.7.0
neo4j